main.sh: check for gpgpass and email files
check for gpgpass and email files before run, as the write script will error out when signing the repo if the files don’t exist

+++ main.sh
@@ -1,13 +1,28 @@
 #!/bin/bash
+GPGPATH="/home/pi/gpgpass"
+EMAILPATH="/home/pi/email"
 git pull origin master
 PKGDIR="/home/pi/raspbian-addons/debian/pool/"
+
+if [ ! -f ${GPGPASS} ]; then
+  echo "error: gpgpass file not detected, exiting."
+  exit 1
+fi
+
+if [ ! -f ${EMAIL} ]; then
+  echo "error: email file not detected, exiting."
+  exit 1
+fi
+
 app="$(echo $1)"
 #echo "enter app to update (use deb package name): "
 #read app
 rm -f $PKGDIR/$app*
 bash packages/${app}.sh
+bash write.sh
 echo "
 Current packages in the repository are:
  "
 ls $PKGDIR
-bash write.sh
+echo "
+Package write for $app complete."
